Refactor project structure: remove main.py, add modular files for hashing and configuration
