+++ .gitignore
@@ -1,0 +1,3 @@
+__pycache__
+*.pyc
+*.pyo
