+++ cynvee_hash_checker/__main__.py
@@ -1,0 +1,19 @@
+"""
+Main entry point for the CynVee-Hash-Checker CLI.
+"""
+
+import sys
+from file_config import file_config
+from path_config import path_config
+from main import main
+
+
+try:
+    if sys.argv[1] == "-n" or sys.argv[1] == "--name":
+        file_config()
+    elif sys.argv[1] == "-p" or sys.argv[1] == "--path":
+        path_config()
+    else:
+        main()
+except IndexError:
+    main()
