+++ pyproject.toml
@@ -1,0 +1,15 @@
+[tool.poetry]
+name = "cynvee_hash_checker"
+version = "0.1.0"
+description = "A project for file configuration and hashing"
+authors = ["Sigireddy Balasai <SigireddyBalasai@gmail.com>"]
+
+[tool.poetry.dependencies]
+python = "^3.8"
+
+[tool.poetry.dev-dependencies]
+pytest = "^6.2"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
